Move LoadComponents to components package
